+++ Helpers/GeneralHelper.cs
@@ -1,6 +1,11 @@
 using AzureNamingTool.Models;
+using AzureNamingTool.Pages;
 using AzureNamingTool.Services;
+using AzureNamingTool.Shared;
+using Blazored.Modal;
+using Blazored.Modal.Services;
 using System.Diagnostics.CodeAnalysis;
+using System.Reflection;
 using System.Security.Cryptography;
 using System.Text;